makedist: update cvmfs package repo urls with fallback
Adds a fallback URL for the cvmfs package download.
makedist currently looks for the latest package in the listing returned by package servers;
since the s3 bucket we are mainly using (and which is the most reliable) does not support listings,
this mechanism now relies on reading a small file .latest_cvmfs_version_prod (https://cvmrepo.web.cern.ch/cvmrepo/.latest_cvmfs_version_prod) containing a string with the latest version.
I'll add this to our testing to ensure it is kept up to date.

Obviously this won't work if you wanted to add mirrors maintained by a third party - one possibility would be to hardcode the latest version in the script.

The downloads for the non-default config packages is kept the same, except for a small fix for the egi download.

The ecsft.cern.ch url is removed alltogether, as this is the most likely one to cause issues.